moved pagination out of data-table.tsx
